chore: million lint setup
