+++ playground/next.config.ts
@@ -1,7 +1,8 @@
 import type { NextConfig } from "next";
+import MillionLint from "@million/lint";
 
 const nextConfig: NextConfig = {
   /* config options here */
 };
 
-export default nextConfig;
+export default MillionLint.next({ rsc: true })(nextConfig);
