+++ playground/package.json
@@ -11,6 +11,7 @@
     "check-types": "tsc --noEmit"
   },
   "dependencies": {
+    "@million/lint": "^1.0.12",
     "@monaco-editor/react": "^4.6.0",
     "@radix-ui/react-icons": "^1.3.1",
     "@radix-ui/react-scroll-area": "^1.2.0",
